third_party/blink/web_tests/platform/generic/fast/js/normal-character-escapes-in-string-literals-expected.txt

Test non-escaping characters in string literals - added with https://bugs.webkit.org/show_bug.cgi?id=100580

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS eval('"\ "') is " "
PASS eval('"\!"') is "!"
PASS eval('"\#"') is "#"
PASS eval('"\$"') is "$"
PASS eval('"\%"') is "%"
PASS eval('"\&"') is "&"
PASS eval('"\'"') is "'"
PASS eval('"\("') is "("
PASS eval('"\)"') is ")"
PASS eval('"\*"') is "*"
PASS eval('"\+"') is "+"
PASS eval('"\,"') is ","
PASS eval('"\-"') is "-"
PASS eval('"\."') is "."
PASS eval('"\/"') is "/"
PASS eval('"\:"') is ":"
PASS eval('"\;"') is ";"
PASS eval('"\<"') is "<"
PASS eval('"\="') is "="
PASS eval('"\>"') is ">"
PASS eval('"\?"') is "?"
PASS eval('"\@"') is "@"
PASS eval('"\A"') is "A"
PASS eval('"\B"') is "B"
PASS eval('"\C"') is "C"
PASS eval('"\D"') is "D"
PASS eval('"\E"') is "E"
PASS eval('"\F"') is "F"
PASS eval('"\G"') is "G"
PASS eval('"\H"') is "H"
PASS eval('"\I"') is "I"
PASS eval('"\J"') is "J"
PASS eval('"\K"') is "K"
PASS eval('"\L"') is "L"
PASS eval('"\M"') is "M"
PASS eval('"\N"') is "N"
PASS eval('"\O"') is "O"
PASS eval('"\P"') is "P"
PASS eval('"\Q"') is "Q"
PASS eval('"\R"') is "R"
PASS eval('"\S"') is "S"
PASS eval('"\T"') is "T"
PASS eval('"\U"') is "U"
PASS eval('"\V"') is "V"
PASS eval('"\W"') is "W"
PASS eval('"\X"') is "X"
PASS eval('"\Y"') is "Y"
PASS eval('"\Z"') is "Z"
PASS eval('"\["') is "["
PASS eval('"\]"') is "]"
PASS eval('"\^"') is "^"
PASS eval('"\_"') is "_"
PASS eval('"\`"') is "`"
PASS eval('"\a"') is "a"
PASS eval('"\c"') is "c"
PASS eval('"\d"') is "d"
PASS eval('"\e"') is "e"
PASS eval('"\g"') is "g"
PASS eval('"\h"') is "h"
PASS eval('"\i"') is "i"
PASS eval('"\j"') is "j"
PASS eval('"\k"') is "k"
PASS eval('"\l"') is "l"
PASS eval('"\m"') is "m"
PASS eval('"\o"') is "o"
PASS eval('"\p"') is "p"
PASS eval('"\q"') is "q"
PASS eval('"\s"') is "s"
PASS eval('"\w"') is "w"
PASS eval('"\y"') is "y"
PASS eval('"\z"') is "z"
PASS eval('"\{"') is "{"
PASS eval('"\|"') is "|"
PASS eval('"\}"') is "}"
PASS eval('"\~"') is "~"
PASS eval('"\"').charCodeAt(0) is 1
PASS eval('"\"').charCodeAt(0) is 7
PASS eval('"\"').charCodeAt(0) is 8
PASS eval('"\"').charCodeAt(0) is 31
PASS eval('"\ "').charCodeAt(0) is 160
PASS eval('"\£"').charCodeAt(0) is 163
PASS eval('"\«"').charCodeAt(0) is 171
PASS eval('"\´"').charCodeAt(0) is 180
PASS successfullyParsed is true

TEST COMPLETE
